Pre-compile dump
Basically rewrote the entire code for visuals, including CSS. Added a carrier module to access the css files and added a designated dark mode file, more themes to come. I also reworked the UI to be more in line with native windows apps and optimized the test we have running in main.